Refactor RuntimeValue utils for re-use
Moving this to tools.parameters.workflow_utils avoids a circular
dependency when importing is_runtime_value in
tools.parameters.dynamic_options.
