--- lib/galaxy/tools/parameters/workflow_building_modes.py
@@ -1,4 +1,0 @@
-class workflow_building_modes:
-    DISABLED = False
-    ENABLED = True
-    USE_HISTORY = 1
